Add titles to player selection
