+++ composeApp/src/commonMain/kotlin/io/music_assistant/client/ui/compose/home/players/SelectPlayerDialog.kt
@@ -1,22 +1,34 @@
 package io.music_assistant.client.ui.compose.home.players
 
+import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.heightIn
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.selection.selectable
 import androidx.compose.foundation.selection.selectableGroup
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.VolumeMute
+import androidx.compose.material.icons.automirrored.filled.VolumeUp
+import androidx.compose.material.icons.filled.Add
 import androidx.compose.material.icons.filled.Link
+import androidx.compose.material.icons.filled.Remove
 import androidx.compose.material3.Card
+import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
 import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.RadioButton
+import androidx.compose.material3.Slider
+import androidx.compose.material3.SliderDefaults
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
@@ -26,8 +38,12 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.semantics.Role
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.DpSize
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Dialog
 import io.music_assistant.client.data.model.client.PlayerData
@@ -41,7 +57,8 @@ fun SelectPlayerDialog(
     players: List<PlayerData>,
     onDismissRequest: () -> Unit,
     onMoveToPlayer: (String) -> Unit = {},
-    groupAction: (String, PlayerAction) -> Unit = { _, _ -> }
+    groupAction: (String, PlayerAction) -> Unit = { _, _ -> },
+    showGroupSettings: Boolean = false
 ) {
     Dialog(onDismissRequest = onDismissRequest) {
         Card(
@@ -49,71 +66,278 @@ fun SelectPlayerDialog(
             shape = RoundedCornerShape(16.dp)
         ) {
             Box(modifier = Modifier.padding(16.dp)) {
-                var showGroupSettings by remember { mutableStateOf(false) }
+                var showGroupSettings by remember { mutableStateOf(showGroupSettings) }
 
-                if (showGroupSettings) {
-                    GroupSettings(
-                        item = selectedPlayer,
-                        onDismissRequest,
-                        groupAction
+                Column {
+                    Text(
+                        if (showGroupSettings) "Group settings" else "Players",
+                        style = MaterialTheme.typography.headlineSmall
                     )
-                } else {
-                    Column(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .selectableGroup()
-                    ) {
-                        players.forEach {
-                            val selected = it.player.id == selectedPlayer.player.id
 
-                            Row(
-                                modifier = Modifier
-                                    .fillMaxWidth()
-                                    .height(56.dp)
-                                    .selectable(
-                                        selected = selected,
-                                        onClick = {
-                                            onDismissRequest()
-                                            onMoveToPlayer(it.player.id)
-                                        },
-                                        role = Role.RadioButton
-                                    ),
-                                verticalAlignment = Alignment.CenterVertically
-                            ) {
-                                RadioButton(
-                                    selected = selected,
-                                    onClick = null
-                                )
-
-                                Text(
-                                    it.player.displayName,
-                                    style = MaterialTheme.typography.bodyLarge,
-                                    modifier = Modifier.padding(start = 16.dp)
-                                )
-                            }
-                        }
-
-                        Row(
-                            modifier = Modifier.fillMaxWidth(),
-                            horizontalArrangement = Arrangement.End
-                        ) {
-                            TextButton(onClick = { showGroupSettings = true }) {
-                                Row(verticalAlignment = Alignment.CenterVertically) {
-                                    Icon(
-                                        imageVector = Icons.Default.Link,
-                                        contentDescription = null
-                                    )
-
-                                    Text(
-                                        "Group",
-                                        modifier = Modifier.padding(start = 4.dp)
-                                    )
-                                }
-                            }
+                    Box(modifier = Modifier.padding(top = 16.dp)) {
+                        if (showGroupSettings) {
+                            GroupSettings(
+                                item = selectedPlayer,
+                                onDismiss = onDismissRequest,
+                                playerAction = groupAction
+                            )
+                        } else {
+                            PlayerSelection(
+                                players,
+                                selectedPlayer,
+                                onDismissRequest,
+                                onMoveToPlayer,
+                                { showGroupSettings = true }
+                            )
                         }
                     }
                 }
             }
+        }
+    }
+}
+
+@Composable
+private fun PlayerSelection(
+    players: List<PlayerData>,
+    selectedPlayer: PlayerData,
+    onDismissRequest: () -> Unit,
+    onSelectPlayer: (String) -> Unit,
+    onShowGroupSettings: () -> Unit
+) {
+    Column(
+        modifier = Modifier
+            .fillMaxWidth()
+            .selectableGroup()
+    ) {
+        players.forEach {
+            val selected = it.player.id == selectedPlayer.player.id
+
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .height(56.dp)
+                    .selectable(
+                        selected = selected,
+                        onClick = {
+                            onDismissRequest()
+                            onSelectPlayer(it.player.id)
+                        },
+                        role = Role.RadioButton
+                    ),
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                RadioButton(
+                    selected = selected,
+                    onClick = null
+                )
+
+                Text(
+                    it.player.displayName,
+                    style = MaterialTheme.typography.bodyLarge,
+                    modifier = Modifier.padding(start = 16.dp)
+                )
+            }
+        }
+
+        Row(
+            modifier = Modifier.fillMaxWidth(),
+            horizontalArrangement = Arrangement.End
+        ) {
+            TextButton(onClick = onShowGroupSettings) {
+                Row(verticalAlignment = Alignment.CenterVertically) {
+                    Icon(
+                        imageVector = Icons.Default.Link,
+                        contentDescription = null
+                    )
+
+                    Text(
+                        "Group",
+                        modifier = Modifier.padding(start = 4.dp)
+                    )
+                }
+            }
+        }
+    }
+}
+
+@Composable
+fun GroupSettings(
+    item: PlayerData,
+    onDismiss: () -> Unit,
+    playerAction: (String, PlayerAction) -> Unit
+) {
+    Column(
+        modifier = Modifier.fillMaxWidth()
+    ) {
+        // Scrollable list of players
+        LazyColumn(
+            modifier = Modifier.fillMaxWidth().heightIn(max = 400.dp),
+            verticalArrangement = Arrangement.spacedBy(8.dp)
+        ) {
+            // Current player at the very top
+            item {
+                GroupPlayerItem(
+                    playerId = item.player.id,
+                    playerName = item.player.name,
+                    isGroup = item.player.isGroup,
+                    volume = if (item.player.isGroup) item.player.groupVolume else item.player.volumeLevel,
+                    isMuted = item.player.volumeMuted.takeIf { item.player.canMute },
+                    simplePlayerAction = playerAction,
+                )
+            }
+
+            // Bound players
+            val boundChildren = item.groupChildren.filter { it.isBound }
+            items(boundChildren, key = { "${it.id}_${it.volume}" }) { child ->
+                GroupPlayerItem(
+                    playerId = child.id,
+                    playerName = child.name,
+                    volume = child.volume,
+                    isMuted = child.isMuted,
+                    simplePlayerAction = playerAction,
+                    bindItem = child,
+                )
+            }
+
+            // Unbound players
+            val unboundChildren = item.groupChildren.filter { !it.isBound }
+            items(unboundChildren, key = { it.id }) { child ->
+                GroupPlayerItem(
+                    playerId = child.id,
+                    playerName = child.name,
+                    volume = child.volume,
+                    isMuted = child.isMuted,
+                    simplePlayerAction = playerAction,
+                    bindItem = child,
+                )
+            }
+        }
+
+        Row(
+            modifier = Modifier.fillMaxWidth(),
+            horizontalArrangement = Arrangement.End
+        ) {
+            TextButton(onClick = onDismiss) {
+                Text("Save")
+            }
+        }
+    }
+}
+
+/**
+ * Group player item with name and volume
+ */
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+private fun GroupPlayerItem(
+    playerId: String,
+    playerName: String,
+    isGroup: Boolean = false,
+    volume: Float?,
+    isMuted: Boolean?,
+    simplePlayerAction: (String, PlayerAction) -> Unit,
+    bindItem: PlayerData.Bind? = null,
+) {
+    var currentVolume by remember(volume) {
+        mutableStateOf(volume ?: 0f)
+    }
+
+    Column(
+        modifier = Modifier.fillMaxWidth(),
+        verticalArrangement = Arrangement.spacedBy((-4).dp)
+    ) {
+        Row(
+            modifier = Modifier.fillMaxWidth(),
+            verticalAlignment = Alignment.CenterVertically,
+            horizontalArrangement = Arrangement.SpaceBetween
+        ) {
+            Text(
+                modifier = Modifier.alpha(if (bindItem?.isBound != false) 1f else 0.4f).weight(1f),
+                text = playerName,
+                maxLines = 1,
+                overflow = TextOverflow.Ellipsis,
+                style = MaterialTheme.typography.bodyMedium,
+                fontWeight = FontWeight.Medium,
+                color = MaterialTheme.colorScheme.onSurface
+            )
+
+            // Show button only for non-current players (when bindItem is provided)
+            bindItem?.let { bind ->
+                val itemId = listOf(playerId)
+                IconButton(
+                    enabled = bindItem.isManageable,
+                    onClick = {
+                        simplePlayerAction(
+                            bind.parentId,
+                            PlayerAction.GroupManage(
+                                toAdd = itemId.takeIf { !bind.isBound },
+                                toRemove = itemId.takeIf { bind.isBound }
+                            )
+                        )
+                    }
+                ) {
+                    Icon(
+                        modifier = Modifier.alpha(if(bindItem.isManageable) 1f else 0.4f),
+                        imageVector = if (bindItem.isBound) Icons.Default.Remove else Icons.Default.Add,
+                        contentDescription = if (bindItem.isBound) "Remove from group" else "Add to group",
+                        tint = if (bindItem.isBound)
+                            MaterialTheme.colorScheme.error
+                        else
+                            MaterialTheme.colorScheme.primary
+                    )
+                }
+            }
+        }
+
+        val volumeEnabled = volume != null && bindItem?.isBound != false
+        Row {
+            isMuted?.let {
+                IconButton(onClick = {
+                    simplePlayerAction(
+                        playerId,
+                        PlayerAction.ToggleMute(isMuted)
+                    )
+                }, enabled = volumeEnabled) {
+                    Icon(
+                        imageVector = if (isMuted) Icons.AutoMirrored.Filled.VolumeMute else Icons.AutoMirrored.Filled.VolumeUp,
+                        contentDescription = if (isMuted) "Unmute" else "Mute"
+                    )
+                }
+            }
+            Slider(
+                modifier = Modifier.fillMaxWidth().alpha(if (volumeEnabled) 1f else 0.4f),
+                value = currentVolume,
+                valueRange = 0f..100f,
+                enabled = volumeEnabled,
+                onValueChange = {
+                    currentVolume = it
+                },
+                onValueChangeFinished = {
+                    simplePlayerAction(
+                        playerId,
+                        if (isGroup) PlayerAction.GroupVolumeSet(currentVolume.toDouble())
+                        else PlayerAction.VolumeSet(currentVolume.toDouble())
+                    )
+                },
+                thumb = {
+                    SliderDefaults.Thumb(
+                        interactionSource = remember { MutableInteractionSource() },
+                        thumbSize = DpSize(16.dp, 16.dp),
+                        colors = SliderDefaults.colors()
+                            .copy(thumbColor = MaterialTheme.colorScheme.secondary),
+                    )
+                },
+                track = { sliderState ->
+                    SliderDefaults.Track(
+                        sliderState = sliderState,
+                        thumbTrackGapSize = 0.dp,
+                        trackInsideCornerSize = 0.dp,
+                        drawStopIndicator = null,
+                        modifier = Modifier.height(4.dp)
+                    )
+                }
+            )
         }
     }
 }
@@ -125,6 +349,18 @@ private fun PreviewSelectPlayerDialog() {
     SelectPlayerDialog(
         selectedPlayer = selectedPlayer,
         players = listOf(selectedPlayer, PlayerDataFixtures.playerData()),
+        onDismissRequest = {}
+    )
+}
+
+@Preview
+@Composable
+private fun PreviewGroupSettingDialog() {
+    val selectedPlayer = PlayerDataFixtures.playerData()
+    SelectPlayerDialog(
+        selectedPlayer = selectedPlayer,
+        players = listOf(selectedPlayer, PlayerDataFixtures.playerData()),
         onDismissRequest = {},
+        showGroupSettings = true,
     )
 }
